@@ -1,5 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
+using System.Security.Cryptography;
 using wan24.Core;
+using wan24.Crypto;
 using wan24.ObjectValidation;
 
 namespace wan24_Crypto_Tests
@@ -16,7 +18,7 @@ namespace wan24_Crypto_Tests
             Logging.Logger = LoggerFactory.CreateLogger("Tests");
             ValidateObject.Logger = (message) => Logging.WriteDebug(message);
             TypeHelper.Instance.ScanAssemblies(typeof(A_Initialization).Assembly);
-            Bootstrap.Async().Wait();
+            wan24.Core.Bootstrap.Async().Wait();
             wan24.Crypto.Bootstrap.Boot();
             DisposableBase.CreateStackInfo = true;
             ErrorHandling.ErrorHandler = (info) =>
@@ -24,6 +26,30 @@ namespace wan24_Crypto_Tests
                 if(info.Exception is StackInfoException six) Logging.WriteError(six.StackInfo.Stack);
             };
             ValidateObject.Logger("wan24-Crypto Tests initialized");
+            // Disable algorithms which are not supported in this platform
+            if (!Shake128.IsSupported)
+            {
+                // SHA3 hash
+                HashHelper.DefaultAlgorithm = HashSha512Algorithm.Instance;
+                HashHelper.Algorithms.TryRemove(HashSha3_256Algorithm.ALGORITHM_NAME, out _);
+                HashHelper.Algorithms.TryRemove(HashSha3_384Algorithm.ALGORITHM_NAME, out _);
+                HashHelper.Algorithms.TryRemove(HashSha3_512Algorithm.ALGORITHM_NAME, out _);
+                HashHelper.Algorithms.TryRemove(HashShake128Algorithm.ALGORITHM_NAME, out _);
+                HashHelper.Algorithms.TryRemove(HashShake256Algorithm.ALGORITHM_NAME, out _);
+                // SHA3 HMAC
+                MacHelper.DefaultAlgorithm = MacHmacSha512Algorithm.Instance;
+                MacHelper.Algorithms.TryRemove(MacHmacSha3_256Algorithm.ALGORITHM_NAME, out _);
+                MacHelper.Algorithms.TryRemove(MacHmacSha3_384Algorithm.ALGORITHM_NAME, out _);
+                MacHelper.Algorithms.TryRemove(MacHmacSha3_512Algorithm.ALGORITHM_NAME, out _);
+                // Pake default options
+                Pake.DefaultOptions = Pake.DefaultOptions
+                    .WithMac(MacHmacSha512Algorithm.ALGORITHM_NAME, included: false);
+                Pake.DefaultCryptoOptions = Pake.DefaultCryptoOptions
+                    .WithMac(MacHmacSha512Algorithm.ALGORITHM_NAME, included: false);
+                // KDF (don't use SHA3 and remove SP800-108)
+                KdfHelper.Algorithms.TryRemove(KdfSp800_108HmacCtrKbKdfAlgorithm.ALGORITHM_NAME, out _);
+                KdfPbKdf2Options.DefaultHashAlgorithm = HashSha384Algorithm.ALGORITHM_NAME;
+            }
         }
     }
 }
