@@ -1,4 +1,5 @@
-﻿using wan24.Crypto;
+﻿using System.Security.Cryptography;
+using wan24.Crypto;
 using wan24.Crypto.Tests;
 
 namespace wan24_Crypto_Tests
@@ -12,7 +13,7 @@ namespace wan24_Crypto_Tests
         [TestMethod]
         public void HashHelper_Tests()
         {
-            Assert.AreEqual(HashSha512Algorithm.ALGORITHM_NAME, HashHelper.DefaultAlgorithm.Name);
+            Assert.AreEqual(SHA3_512.IsSupported ? HashSha3_512Algorithm.ALGORITHM_NAME : HashSha512Algorithm.ALGORITHM_NAME, HashHelper.DefaultAlgorithm.Name);
             Assert.AreEqual(HashHelper.DefaultAlgorithm.HashLength, TestData.Data.Hash().Length);
             Assert.AreEqual(HashHelper.DefaultAlgorithm, HashHelper.GetAlgorithm(HashHelper.DefaultAlgorithm.Name));
             Assert.AreEqual(HashHelper.DefaultAlgorithm, HashHelper.GetAlgorithm(HashHelper.DefaultAlgorithm.Value));
