@@ -1,5 +1,4 @@
-﻿using System.Security.Cryptography;
-using wan24.Core;
+﻿using wan24.Core;
 using wan24.Crypto;
 using wan24.Crypto.Tests;
 
@@ -24,7 +23,6 @@ namespace wan24_Crypto_Tests
         public void Options_Tests()
         {
             KdfPbKdf2Options kdfOptions = new();
-            Assert.AreEqual(HashAlgorithmName.SHA384, kdfOptions.HashName);
             CryptoOptions options = new();
             options.WithKdf(KdfPbKdf2Algorithm.ALGORITHM_NAME, kdfOptions: kdfOptions);
             KdfHelper.Stretch("test".GetBytes(), len: 32, options: options);
