@@ -1,4 +1,5 @@
-﻿using wan24.Crypto;
+﻿using System.Security.Cryptography;
+using wan24.Crypto;
 using wan24.Crypto.Tests;
 
 namespace wan24_Crypto_Tests
@@ -12,7 +13,7 @@ namespace wan24_Crypto_Tests
         [TestMethod]
         public void MacHelper_Tests()
         {
-            Assert.AreEqual(MacHmacSha512Algorithm.ALGORITHM_NAME, MacHelper.DefaultAlgorithm.Name);
+            Assert.AreEqual(HMACSHA3_512.IsSupported ? MacHmacSha3_512Algorithm.ALGORITHM_NAME : MacHmacSha512Algorithm.ALGORITHM_NAME, MacHelper.DefaultAlgorithm.Name);
             Assert.AreEqual(MacHelper.DefaultAlgorithm.MacLength, TestData.Data.Hash().Length);
             Assert.AreEqual(MacHelper.DefaultAlgorithm, MacHelper.GetAlgorithm(MacHelper.DefaultAlgorithm.Name));
             Assert.AreEqual(MacHelper.DefaultAlgorithm, MacHelper.GetAlgorithm(MacHelper.DefaultAlgorithm.Value));
