@@ -1,4 +1,6 @@
-﻿namespace wan24.Crypto
+﻿using wan24.Core;
+
+namespace wan24.Crypto
 {
     /// <summary>
     /// Asymmetric algorithm usages
@@ -9,10 +11,12 @@
         /// <summary>
         /// Key exchange
         /// </summary>
+        [DisplayText("Key exchange")]
         KeyExchange = 1,
         /// <summary>
         /// Signature
         /// </summary>
+        [DisplayText("Signature")]
         Signature = 2
     }
 }
