@@ -24,9 +24,10 @@ namespace wan24.Crypto
         {
             Algorithms = new(new KeyValuePair<string, KdfAlgorithmBase>[]
             {
-                new(KdfPbKdf2Algorithm.ALGORITHM_NAME, KdfPbKdf2Algorithm.Instance)
+                new(KdfPbKdf2Algorithm.ALGORITHM_NAME, KdfPbKdf2Algorithm.Instance),
+                new(KdfSp800_108HmacCtrKbKdfAlgorithm.ALGORITHM_NAME, KdfSp800_108HmacCtrKbKdfAlgorithm.Instance)
             });
-            _DefaultAlgorithm = Algorithms[KdfPbKdf2Algorithm.ALGORITHM_NAME];
+            _DefaultAlgorithm = KdfPbKdf2Algorithm.Instance;
         }
 
         /// <summary>
