@@ -126,7 +126,7 @@ namespace wan24.Crypto
                 if (CryptoHelper.StrictPostQuantumSafety && !IsPostQuantum) throw new InvalidOperationException($"Post quantum safety-forced - {Name} isn't post quantum");
                 options = MacHelper.GetDefaultOptions(options?.GetCopy() ?? DefaultOptions);
                 byte[] res = new byte[MacLength];
-                GetMacAlgorithm(pwd, options).TryComputeHash(data, res, out _);
+                if (!GetMacAlgorithm(pwd, options).TryComputeHash(data, res, out _)) throw new IOException($"Failed to compute the final MAC");
                 return res;
             }
             catch (CryptographicException)
@@ -154,7 +154,7 @@ namespace wan24.Crypto
                 if (outputBuffer.Length < MacLength) throw new ArgumentOutOfRangeException(nameof(outputBuffer));
                 if (CryptoHelper.StrictPostQuantumSafety && !IsPostQuantum) throw new InvalidOperationException($"Post quantum safety-forced - {Name} isn't post quantum");
                 options = MacHelper.GetDefaultOptions(options?.GetCopy() ?? DefaultOptions);
-                GetMacAlgorithm(pwd, options).TryComputeHash(data, outputBuffer, out _);
+                if (!GetMacAlgorithm(pwd, options).TryComputeHash(data, outputBuffer, out _)) throw new IOException($"Failed to compute the final MAC");
                 return outputBuffer;
             }
             catch (CryptographicException)
